feat(workflow-designer-ui): integrate InputPanel for image and text generation nodes
